Move temperature convert to seperate module
